backend/requirements.txt
tqdm
numpy
torch>=2.2.0
transformers>=4.43.0
datasets
einops>=0.8.0
tiktoken==0.7.0
accelerate==0.33.0
fastapi==0.112.1
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==23.2.1